refactor(fuji): flatten workspace/ into single workspace.ts
Merge definition.ts and workspace.ts into lib/workspace.ts. The two-file
directory added indirection for one table and one action. Update all
imports from $lib/workspace/definition and ./workspace/workspace.
